- added about component - added toastr functionality to article details - added logo and favicon

+++ Awesome-News-System-4000/app/article-details/article-details.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
+import { ToastsManager } from 'ng2-toastr/ng2-toastr';
+
 import { ArticleDetailsModel } from '../../app/core/models/article-details.model';
 import { ArticleDetailsService } from './article-details.service';
 import { AuthenticationService } from '../core/services/authentication.service';
@@ -22,7 +24,8 @@ export class ArticleDetailsComponent implements OnInit {
 
     constructor(private _service: ArticleDetailsService,
         private _authenticationService: AuthenticationService,
-        private _route: ActivatedRoute) {
+        private _route: ActivatedRoute,
+        public toastr: ToastsManager) {
     }
     ngOnInit() {
         this.getArticleDetails(this._route.snapshot.params['id']);
@@ -40,35 +43,55 @@ export class ArticleDetailsComponent implements OnInit {
             .subscribe(response => {
                 this.article.rating = response.article.rating;
                 this.isRatingAdded = true;
+                this.toastr.success('Rating successfully submited!');
             });
     }
     onAddToFavourites() {
-        this._service.addArticleToFavourites(this.article._id, this.user, this._route.snapshot.params['id'])
-            .subscribe(response => {
-                this.isAddedToFavourites = true;
-                let updatedUser = JSON.parse(this._authenticationService.checkIfUserIsLoggedIn()).user;
-                updatedUser.favouriteArticles.push({
-                    _id: this.article._id,
-                    imageUrl: this.article.imageUrl,
-                    originalId: this._route.snapshot.params['id'],
-                    publishedAt: this.article.publishedAt,
-                    source: this.article.source,
-                    title: this.article.title
+        if (!this.isAlreadyInFavourites()) {
+            this._service.addArticleToFavourites(this.article._id, this.user, this._route.snapshot.params['id'])
+                .subscribe(response => {
+                    this.isAddedToFavourites = true;
+                    let updatedUser = JSON.parse(this._authenticationService.checkIfUserIsLoggedIn()).user;
+                    updatedUser.favouriteArticles.push({
+                        _id: this.article._id,
+                        imageUrl: this.article.imageUrl,
+                        originalId: this._route.snapshot.params['id'],
+                        publishedAt: this.article.publishedAt,
+                        source: this.article.source,
+                        title: this.article.title
+                    });
+                    localStorage.removeItem('currentUser');
+                    localStorage.setItem('currentUser', JSON.stringify({ user: updatedUser }));
+                    this.toastr.success('Article added to favourites list!');
+                }, error => {
+                    this.toastr.error('Article already in favourites.');
                 });
-                localStorage.removeItem('currentUser');
-                localStorage.setItem('currentUser', JSON.stringify({ user: updatedUser }));
-            }, error => {
-                alert('Item already in favourites.');
-            });
+        } else {
+            this.toastr.error('Article is already in favourites.');
+        }
     }
     onAddCommentToArticle() {
         this._service.addCommentToArticle(this.articleCommentToAdd, this.article._id, this.user).subscribe(
-            item => this.article.comments = item.comments
-        );
+            item => {
+                this.article.comments = item.comments;
+                this.toastr.success('Comment added successfully!');
+            });
     }
-    getArticleDetails(id: string) {
+    private getArticleDetails(id: string) {
         this._service.getArticle(id).subscribe(
             data => this.article = data.result
         );
+    }
+    private isAlreadyInFavourites(): boolean {
+        let currentId = this._route.snapshot.params['id'];
+        let userLocalStorage = JSON.parse(this._authenticationService.checkIfUserIsLoggedIn()).user;
+
+        for (let i = 0; i < userLocalStorage.favouriteArticles.length; i++) {
+            if (currentId === userLocalStorage.favouriteArticles[i].originalId) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
